Fix for neutral alignment summon lesser servitor (#93)
My previous suggested fix for neutral summon lesser servitor didn't seem
to work. I've tested this fix and it works correctly for neutral and
non-neutral alignments. The issue was that when a neutral alignment was
reversed the alignmentStringToTraits function returned an empty string
which cause all of the creatures to be filtered out (the some function
was always returning false). I added an additional conditional to ignore
the filter if the reverse alignment is an empty string.